GhePlots

Documentation for GhePlots.

GhePlots.dark_mode

GhePlots.light_mode

GhePlots.siground

GhePlots.sub

GhePlots.sup

GhePlots.dark_mode — Method

dark_mode()

Makie theme for dark backgrounds.

GhePlots.light_mode — Method

light_mode()

Makie theme for white backgrounds.

GhePlots.siground — Method

siground(x,u,sigfigs=3)

Round a measurement x and its uncertainty u to sigfigs. Accepts a single value or a Tuple of values for u.

Returns a Tuple of (x,u...)

e.g. julia> siground(523.12, (15.8,278.0) ) (523, 16, 278)

GhePlots.sub — Method

sub(I)

Return a unicode string subscript of Integer I.

GhePlots.sup — Method

sup(I)

Return a unicode string superscript of Integer I.